Updated the testing infra
The new infrasctructure allows for tests to be run individually, thus
allowing us to extract a global % of coverage from running all tests.

I am using an `(or t)` selector to run tests because of a UI bug in DOOM
Emacs which does not allow me to enter `t` to run all tests.